Delete unused harness adapter classes
